Move from . -> src
