+++ src/table2xsv/__init__.py
@@ -1,3 +1,3 @@
 __author__ = "Amogh Madan"
 __name__ = "Table2XSV"
-__version__ = "2.0.1"
+__version__ = "2.2.0"
