+++ src/table2xsv/core/__init__.py
@@ -1,9 +1,9 @@
-from table2xsv.core.base import BaseCommand
+from table2xsv.core.extended import Table2XSVBaseCommand
 from table2xsv.core.exceptions import CommandError, SystemCheckError
 from table2xsv.core.utility import ManagementUtility
 
 __all__ = [
-    "BaseCommand",
+    "Table2XSVBaseCommand",
     "CommandError",
     "SystemCheckError",
     "ManagementUtility",
