+++ src/table2xsv/core/exceptions.py
