+++ src/table2xsv/core/formatter.py
