+++ src/table2xsv/core/helpers.py
