+++ src/table2xsv/core/parser.py
