+++ src/table2xsv/core/utility.py
