+++ src/table2xsv/core/wrapper.py
