+++ src/table2xsv/utils/getter.py
