--- table2xsv/management/__init__.py
@@ -1,5 +1,0 @@
-from table2xsv.management.base import Table2XSVBaseCommand
-
-__all__ = [
-    "Table2XSVBaseCommand",
-]
